Event name refactor
