+++ src/main/java/org/teleight/teleightbots/event/bot/group/BotJoinedGroupEvent.java
@@ -5,7 +5,7 @@ import org.teleight.teleightbots.api.objects.Update;
 import org.teleight.teleightbots.bot.Bot;
 import org.teleight.teleightbots.event.trait.GroupBotEvent;
 
-public record BotJoinGroupEvent(
+public record BotJoinedGroupEvent(
         @NotNull Bot bot,
         @NotNull Update update
 ) implements GroupBotEvent {
