+++ src/main/java/org/teleight/teleightbots/event/user/UserInlineQueryReceivedEvent.java
@@ -1,23 +1,24 @@
 package org.teleight.teleightbots.event.user;
 
 import org.jetbrains.annotations.NotNull;
-import org.teleight.teleightbots.api.objects.Message;
 import org.teleight.teleightbots.api.objects.Update;
+import org.teleight.teleightbots.api.objects.User;
+import org.teleight.teleightbots.api.objects.inline.InlineQuery;
 import org.teleight.teleightbots.bot.Bot;
 import org.teleight.teleightbots.event.trait.Event;
 
-public record UserWriteEvent(
+public record UserInlineQueryReceivedEvent(
         @NotNull Bot bot,
         @NotNull Update update
 ) implements Event {
 
-    @Override
-    public @NotNull Update update() {
-        return update;
+    @NotNull
+    public User user() {
+        return update().inlineQuery().from();
     }
 
-    public Message message(){
-        return update.message();
+    public InlineQuery inlineQuery() {
+        return update().inlineQuery();
     }
 
 }
