+++ src/main/java/org/teleight/teleightbots/event/user/UserMessageReceivedEvent.java
@@ -1,23 +1,23 @@
 package org.teleight.teleightbots.event.user;
 
 import org.jetbrains.annotations.NotNull;
+import org.teleight.teleightbots.api.objects.Message;
 import org.teleight.teleightbots.api.objects.Update;
-import org.teleight.teleightbots.api.objects.User;
-import org.teleight.teleightbots.api.objects.inline.InlineQuery;
 import org.teleight.teleightbots.bot.Bot;
 import org.teleight.teleightbots.event.trait.Event;
 
-public record UserInlineSentEvent(
+public record UserMessageReceivedEvent(
         @NotNull Bot bot,
         @NotNull Update update
 ) implements Event {
 
-    public User user() {
-        return update().inlineQuery().from();
+    @Override
+    public @NotNull Update update() {
+        return update;
     }
 
-    public InlineQuery inlineQuery() {
-        return update().inlineQuery();
+    public Message message(){
+        return update.message();
     }
 
 }
